requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
mediapipe==0.10.14
opencv-python==4.10.0.82
numpy==1.26.4
python-multipart==0.0.9
pytest==8.2.1
httpx==0.27.0
Pillow==10.3.0
